// src/types.ts
export interface InfactorySDKError {
  status: number;
  message: string;
}
export interface ApiResponse<T> {
  data?: T;
  error?: InfactorySDKError;
}

// Base types
export interface BaseEntity {
  id: string;
  created_at: string;
  updated_at: string;
  deleted_at?: string | null;
}
// ...
export interface Project extends BaseEntity {
  name: string;
  description?: string;
  team_id?: string;
}

export interface CreateProjectParams {
  name: string;
  description?: string;
  team_id: string;
}